+++ packages/angular/src/lib/chat.ng.test.ts
@@ -416,6 +416,7 @@ describe('onToolCall', () => {
           input: { testArg: 'test-value' },
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -434,6 +435,7 @@ describe('onToolCall', () => {
         output:
           'test-tool-response: test-tool tool-call-0 {"testArg":"test-value"}',
         providerExecuted: undefined,
+        preliminary: undefined,
       },
     ]);
   });
@@ -483,6 +485,7 @@ describe('tool invocations', () => {
           input: undefined,
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -506,6 +509,7 @@ describe('tool invocations', () => {
           input: { testArg: 't' },
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -529,6 +533,7 @@ describe('tool invocations', () => {
           input: { testArg: 'test-value' },
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -553,6 +558,7 @@ describe('tool invocations', () => {
           input: { testArg: 'test-value' },
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -577,6 +583,7 @@ describe('tool invocations', () => {
         input: { testArg: 'test-value' },
         output: 'test-result',
         providerExecuted: undefined,
+        preliminary: undefined,
       },
     ]);
   });
@@ -610,6 +617,7 @@ describe('tool invocations', () => {
           input: { testArg: 'test-value' },
           output: undefined,
           providerExecuted: undefined,
+          preliminary: undefined,
         },
       ]);
     });
@@ -635,6 +643,7 @@ describe('tool invocations', () => {
         input: { testArg: 'test-value' },
         output: 'test-result',
         providerExecuted: undefined,
+        preliminary: undefined,
       },
     ]);
   });
@@ -666,6 +675,7 @@ describe('tool invocations', () => {
           type: 'tool-test-tool',
           input: { testArg: 'test-value' },
           output: undefined,
+          preliminary: undefined,
           providerExecuted: undefined,
         },
       ]);
@@ -687,6 +697,7 @@ describe('tool invocations', () => {
           type: 'tool-test-tool',
           input: { testArg: 'test-value' },
           output: 'test-result',
+          preliminary: undefined,
           providerExecuted: undefined,
         },
       ]);