+++ packages/elevenlabs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/elevenlabs
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [034e229]
+- Updated dependencies [f25040d]
+  - @ai-sdk/provider-utils@3.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [38ac190]
+  - @ai-sdk/provider-utils@3.0.2
+
 ## 1.0.1
 
 ### Patch Changes