+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/groq
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [034e229]
+- Updated dependencies [f25040d]
+  - @ai-sdk/provider-utils@3.0.3
+
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [38ac190]
+  - @ai-sdk/provider-utils@3.0.2
+
 ## 2.0.5
 
 ### Patch Changes